# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "gin-example",
        "contact": {
            "name": "guanwei",
            "email": "riverhohai@gmail.com"
        },
        "license": {}
    },
    "paths": {
        "/vpi/v1/articles": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "添加新的文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TagId",
                        "name": "tag_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Desc",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Content",
                        "name": "content",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "CreatedBy",
                        "name": "created_by",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"msg\": \"ok\", \"data\": null \"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}